requirements.txt
absl-py==2.1.0
accelerate==0.33.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
appdirs==1.4.4
asttokens==2.4.1
attrs==24.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
bitsandbytes==0.42.0
black==24.8.0
blinker==1.8.2
boltons==21.0.0
bracex==2.5
Brotli==1.1.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-option-group==0.5.6
codeshield==1.0.1
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.0
cryptography==43.0.0
cycler==0.12.1
dataclasses-json==0.6.7
datasets==2.21.0
decorator==5.1.1
deepdiff==8.0.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.8
distro==1.9.0
dnspython==2.6.1
effdet==0.4.1
emoji==2.12.1
evaluate==0.4.2
exceptiongroup==1.2.2
executing==2.0.1
face==22.0.0
fastapi==0.112.2
ffmpy==0.4.0
filelock==3.15.4
filetype==1.2.0
fire==0.6.0
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
glom==22.1.0
google-api-core==2.19.2
google-auth==2.34.0
google-cloud-vision==3.7.4
googleapis-common-protos==1.65.0
gradio==4.42.0
gradio_client==1.3.0
grpcio==1.66.1
grpcio-status==1.62.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.24.6
humanfriendly==10.0
idna==3.8
importlib_metadata==7.1.0
importlib_resources==6.4.4
inflate64==1.0.0
iopath==0.1.10
ipython==8.26.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
langchain==0.2.16
langchain-community==0.2.17
langchain-core==0.2.40
langchain-openai==0.1.20
langchain-text-splitters==0.2.2
langdetect==1.0.9
langsmith==0.1.120
layoutparser==0.3.4
loralib==0.1.2
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
motor==3.5.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
multivolumefile==0.2.3
mypy-extensions==1.0.0
narwhals==1.8.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
numexpr==2.10.1
numpy==1.26.4
omegaconf==2.3.0
onnx==1.16.2
onnxruntime==1.19.0
openai==1.39.0
opencv-python==4.10.0.84
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-http==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-requests==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0
optimum==1.21.4
orderly-set==5.2.2
orjson==3.10.7
packaging==24.1
pandas==2.2.2
parso==0.8.4
pathspec==0.12.1
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.4
peewee==3.17.6
peft==0.12.0
pexpect==4.9.0
pi_heif==0.18.0
pikepdf==9.2.0
pillow==10.4.0
platformdirs==4.2.2
portalocker==2.10.1
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.4
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py7zr==0.22.0
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pybcj==1.0.2
pycocotools==2.0.8
pycparser==2.22
pycryptodomex==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
pymongo==4.8.0
pyparsing==3.1.4
pypdf==4.3.1
pypdfium2==4.30.0
pyppmd==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-iso639==2024.4.27
python-magic==0.4.27
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
pyzstd==0.16.1
rapidfuzz==3.9.6
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.8.0
rouge_score==0.1.2
rpds-py==0.20.0
rsa==4.9
ruamel.yaml==0.17.40
ruamel.yaml.clib==0.2.8
ruff==0.6.3
safetensors==0.4.4
scikit-learn==1.5.1
scipy==1.14.1
semantic-version==2.10.0
semgrep==1.85.0
sentence-transformers==3.0.1
sentencepiece==0.2.0
setuptools==74.0.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.32
stack-data==0.6.3
starlette==0.38.2
streamlit==1.38.0
sympy==1.13.2
tabulate==0.9.0
tenacity==8.5.0
termcolor==2.4.0
texttable==1.7.0
threadpoolctl==3.5.0
tiktoken==0.7.0
timm==1.0.9
tokenize-rt==6.0.0
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.0
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.43.4
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
unstructured==0.15.8
unstructured-client==0.25.5
unstructured-inference==0.7.36
unstructured.pytesseract==0.3.13
urllib3==2.2.2
uvicorn==0.30.6
watchdog==5.0.2
wcmatch==8.5.2
wcwidth==0.2.13
websockets==12.0
wikipedia==1.4.0
wrapt==1.16.0
xxhash==3.5.0
yarl==1.9.4
zipp==3.20.1
